+++ cvat-ui/src/components/projects-page/project-list.tsx
@@ -1,4 +1,4 @@
-// Copyright (C) 2020 Intel Corporation
+// Copyright (C) 2020-2021 Intel Corporation
 //
 // SPDX-License-Identifier: MIT
 
@@ -8,14 +8,14 @@ import { Row, Col } from 'antd/lib/grid';
 import Pagination from 'antd/lib/pagination';
 
 import { getProjectsAsync } from 'actions/projects-actions';
-import { CombinedState } from 'reducers/interfaces';
+import { CombinedState, Project } from 'reducers/interfaces';
 import ProjectItem from './project-item';
 
 export default function ProjectListComponent(): JSX.Element {
     const dispatch = useDispatch();
     const projectsCount = useSelector((state: CombinedState) => state.projects.count);
     const { page } = useSelector((state: CombinedState) => state.projects.gettingQuery);
-    let projectInstances = useSelector((state: CombinedState) => state.projects.current);
+    const projectInstances = useSelector((state: CombinedState) => state.projects.current);
     const gettingQuery = useSelector((state: CombinedState) => state.projects.gettingQuery);
 
     function changePage(p: number): void {
@@ -27,7 +27,7 @@ export default function ProjectListComponent(): JSX.Element {
         );
     }
 
-    projectInstances = projectInstances.reduce((rows, key, index) => {
+    const projects = projectInstances.reduce<Project[][]>((rows, key, index) => {
         if (index % 4 === 0) {
             rows.push([key]);
         } else {
@@ -38,14 +38,14 @@ export default function ProjectListComponent(): JSX.Element {
 
     return (
         <>
-            <Row justify='center' align='middle'>
+            <Row justify='center' align='middle' className='cvat-project-list-content'>
                 <Col className='cvat-projects-list' md={22} lg={18} xl={16} xxl={14}>
-                    {projectInstances.map(
-                        (row: any[]): JSX.Element => (
-                            <Row key={row[0].id} gutter={[8, 8]}>
-                                {row.map((instance: any) => (
-                                    <Col span={6} key={instance.id}>
-                                        <ProjectItem projectInstance={instance} />
+                    {projects.map(
+                        (row: Project[]): JSX.Element => (
+                            <Row key={row[0].instance.id} gutter={[8, 8]}>
+                                {row.map((project: Project) => (
+                                    <Col span={6} key={project.instance.id}>
+                                        <ProjectItem projectInstance={project} />
                                     </Col>
                                 ))}
                             </Row>